backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
tensorflow
opencv-python
numpy
Pillow
python-multipart
